+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Support functional engine ([#42](https://github.com/marp-team/marp-cli/pull/42))
+- Output the configured engine in `version` (`-v`) option ([#43](https://github.com/marp-team/marp-cli/pull/43))
 
 ## v0.0.14 - 2018-11-24